+++ R/expect-equality.R
@@ -4,6 +4,8 @@
 #' - `expect_equal` tests with [all.equal()]
 #' - `expect_equivalent` tests with [all.equal()] and
 #'   `check.attributes = FALSE`
+#' - `expect_reference` tests if two symbols point to the same underlying
+#'    object in memory
 #
 #' @param expected Expected value
 #' @param expected.label Equivalent of `label` for shortcut form.
@@ -103,3 +105,46 @@ expect_identical <- function(object, expected, info = NULL, label = NULL,
   invisible(object)
 }
 
+#' @export
+#' @rdname equality-expectations
+expect_identical <- function(object, expected, info = NULL, label = NULL,
+                             expected.label = NULL) {
+
+  lab_act <- make_label(object, label)
+  lab_exp <- make_label(expected, expected.label)
+
+  ident <- identical(object, expected)
+  if (ident) {
+    msg <- ""
+  } else {
+    compare <- compare(object, expected)
+    if (compare$equal) {
+      msg <- "Objects equal but not identical"
+    } else {
+      msg <- compare$message
+    }
+  }
+
+  expect(
+    ident,
+    sprintf("%s not identical to %s.\n%s", lab_act, lab_exp, msg),
+    info = info
+  )
+  invisible(object)
+}
+
+#' @export
+#' @rdname equality-expectations
+expect_reference <- function(object, expected, info = NULL, label = NULL,
+                             expected.label = NULL) {
+
+  lab_act <- make_label(object, label)
+  lab_exp <- make_label(expected, expected.label)
+
+  expect(
+    is_reference(object, expected),
+    sprintf("%s not a reference %s.", lab_act, lab_exp),
+    info = info
+  )
+  invisible(object)
+}