+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `with_mock()` disallows mocking of functions in base packages, because this doesn't work with the current development version of R (#553).
 
+* `expect_reference()` checks if two names point to the same object (#622).
+
 * Output expectations (`expect_output()`, `expect_message()`, 
   `expect_warning()`, and `expect_silent()`) all invisibly return the first
   argument to be consistent with the other expectations (#615).